requirements.txt
Django==2.0
django-bootstrap-form==3.3
olefile==0.44
Pillow==4.3.0
PyJWT==1.5.3
pytz==2017.3
